move the doctest data into a data package
